clean userV2, decouple optional user info, revert test fixes
