+++ src/main/OptionalUserInformation/MaritalStatus.java
@@ -1,4 +1,4 @@
-package UserV2;
+package OptionalUserInformation;
 
 public enum MaritalStatus {
   SINGLE,
