+++ src/main/OptionalUserInformation/Person.java
@@ -1,4 +1,4 @@
-package UserV2;
+package OptionalUserInformation;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
