+++ src/main/OptionalUserInformation/Services/UpdateOptionalInfoService.java
@@ -1,9 +1,9 @@
-package UserV2.Services;
+package OptionalUserInformation.Services;
 
 import Config.Message;
 import Config.Service;
 import Database.OptionalUserInformation.OptionalUserInformationDao;
-import UserV2.*;
+import OptionalUserInformation.*;
 
 public class UpdateOptionalInfoService implements Service {
     OptionalUserInformationDao optionalUserInformationDao;
