--- src/main/UserV2/IpObject.java
@@ -1,46 +1,0 @@
-package UserV2;
-
-public class IpObject {
-
-  private String ip;
-
-  private String location;
-
-  private String date;
-
-  private String device;
-
-  public IpObject() {}
-
-  public String getDevice() {
-    return device;
-  }
-
-  public void setDevice(String device) {
-    this.device = device;
-  }
-
-  public String getDate() {
-    return date;
-  }
-
-  public void setDate(String date) {
-    this.date = date;
-  }
-
-  public String getLocation() {
-    return location;
-  }
-
-  public void setLocation(String location) {
-    this.location = location;
-  }
-
-  public String getIp() {
-    return ip;
-  }
-
-  public void setIp(String ip) {
-    this.ip = ip;
-  }
-}
