--- src/main/UserV2/VeteranStatus.java
@@ -1,27 +1,0 @@
-package UserV2;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.*;
-
-import java.util.Map;
-
-@Getter
-@Setter
-@Builder(toBuilder = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class VeteranStatus {
-
-  @NonNull private boolean isVeteran;
-  @NonNull private boolean isProtectedVeteran;
-  private String branch;
-  private String yearsOfService;
-  private String rank;
-  private String discharge;
-
-  public Map<String, Object> toMap() {
-    ObjectMapper objectMapper = new ObjectMapper();
-    return objectMapper.convertValue(this, new TypeReference<>() {});
-  }
-}
