+++ src/test/OptionalUserInformationTest/CreateOptionalUserInformationUnitTest.java
@@ -4,8 +4,8 @@ import Config.DeploymentLevel;
 import Config.Message;
 import Database.OptionalUserInformation.OptionalUserInformationDao;
 import Database.OptionalUserInformation.OptionalUserInformationDaoFactory;
-import UserV2.*;
-import UserV2.Services.CreateOptionalInfoService;
+import OptionalUserInformation.*;
+import OptionalUserInformation.Services.CreateOptionalInfoService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.After;
 import org.junit.Test;
