+++ src/test/OptionalUserInformationTest/GetOptionalUserInformationUnitTest.java
@@ -4,9 +4,9 @@ import Config.DeploymentLevel;
 import Config.Message;
 import Database.OptionalUserInformation.OptionalUserInformationDao;
 import Database.OptionalUserInformation.OptionalUserInformationDaoFactory;
-import UserV2.*;
-import UserV2.Services.CreateOptionalInfoService;
-import UserV2.Services.GetOptionalInfoService;
+import OptionalUserInformation.*;
+import OptionalUserInformation.Services.CreateOptionalInfoService;
+import OptionalUserInformation.Services.GetOptionalInfoService;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
 import org.junit.After;
