+++ src/test/OptionalUserInformationTest/UpdateOptionalUserInformationUnitTest.java
@@ -4,9 +4,9 @@ import Config.DeploymentLevel;
 import Config.Message;
 import Database.OptionalUserInformation.OptionalUserInformationDao;
 import Database.OptionalUserInformation.OptionalUserInformationDaoFactory;
-import UserV2.*;
-import UserV2.Services.CreateOptionalInfoService;
-import UserV2.Services.UpdateOptionalInfoService;
+import OptionalUserInformation.*;
+import OptionalUserInformation.Services.CreateOptionalInfoService;
+import OptionalUserInformation.Services.UpdateOptionalInfoService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.After;
 import org.junit.Test;
